Refactored some old files
DimManager is now WorldNumbered. Removed some unused methods
