+++ Claims/src/net/pgfmc/claims/ownable/block/events/BlockInteractEvent.java
@@ -15,7 +15,7 @@ import net.pgfmc.claims.ownable.block.ClaimConfigInventory;
 import net.pgfmc.claims.ownable.block.table.ClaimsLogic.Range;
 import net.pgfmc.claims.ownable.block.table.ClaimsTable;
 import net.pgfmc.core.api.playerdata.PlayerData;
-import net.pgfmc.core.util.Vector4;
+import net.pgfmc.core.util.vector4.Vector4;
 
 /**
 Written by CrimsonDart
