+++ Claims/src/net/pgfmc/claims/ownable/block/events/ItemFrameBreak.java
@@ -12,7 +12,7 @@ import net.pgfmc.claims.ownable.block.Claim.Security;
 import net.pgfmc.claims.ownable.block.table.ClaimsLogic.Range;
 import net.pgfmc.claims.ownable.block.table.ClaimsTable;
 import net.pgfmc.core.api.playerdata.PlayerData;
-import net.pgfmc.core.util.Vector4;
+import net.pgfmc.core.util.vector4.Vector4;
 
 public class ItemFrameBreak implements Listener {
