+++ Claims/src/net/pgfmc/claims/ownable/block/table/ClaimSection.java
@@ -8,7 +8,7 @@ import java.util.Set;
 import net.pgfmc.claims.ownable.block.Claim;
 import net.pgfmc.claims.ownable.block.table.ClaimsLogic.Range;
 import net.pgfmc.claims.util.LongHash;
-import net.pgfmc.core.util.Vector4;
+import net.pgfmc.core.util.vector4.Vector4;
 
 /**
  * Stores claims in a certain region.
