+++ Claims/src/net/pgfmc/claims/ownable/inspector/ClaimTPCommand.java
@@ -8,8 +8,8 @@ import net.pgfmc.claims.ownable.block.Claim;
 import net.pgfmc.claims.ownable.block.table.ClaimsLogic.Range;
 import net.pgfmc.claims.ownable.block.table.ClaimsTable;
 import net.pgfmc.core.api.playerdata.PlayerData;
-import net.pgfmc.core.util.Vector4;
 import net.pgfmc.core.util.commands.PlayerCommand;
+import net.pgfmc.core.util.vector4.Vector4;
 
 public class ClaimTPCommand extends PlayerCommand {
