+++ Core/src/net/pgfmc/core/api/teleport/SafeLocation.java
@@ -9,7 +9,7 @@ import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 
-import net.pgfmc.core.util.Vector4;
+import net.pgfmc.core.util.vector4.Vector4;
 
 public class SafeLocation implements Runnable {
